refactor isValid for better dx
