+++ src/components/Event/EventFields/styles.module.scss
@@ -8,29 +8,6 @@
     justify-content: center;
     display: flex;
 }
-.conflicts {
-    position: relative;
-    .overlappingEvents {
-        display: flex;
-        flex-wrap: wrap;
-        gap: 0.5em;
-        max-height: 30em;
-        overflow-y: auto;
-        .overview {
-            flex-shrink: 1;
-            flex-basis: 15em;
-        }
-        .main {
-            flex-shrink: 0;
-            flex-grow: 1;
-            flex-basis: 100%;
-            border-width: 4px;
-            position: sticky;
-            top: 0;
-            background: var(--ifm-navbar-background-color);
-        }
-    }
-}
 .errors {
     color: var(--ifm-color-danger);
 }
