+++ src/components/Event/EventOverviewSmall/styles.module.scss
@@ -9,7 +9,10 @@
     border: 1px solid;
     padding: 2px 4px;
     border-radius: var(--ifm-global-radius);
-
+    cursor: pointer;
+    &:hover {
+        box-shadow: var(--ifm-global-shadow-md);
+    }
     .header {
         grid-area: header;
         display: grid;
@@ -21,11 +24,10 @@
         font-size: 10px;
         .location {
             grid-area: location;
-
             justify-self: center;
         }
         .date {
-            font-weight: 600;
+            font-weight: 700;
             &.dateStart {
                 grid-area: d-start;
             }
@@ -49,11 +51,11 @@
     .body {
         grid-area: body;
         overflow: hidden;
-        .title {
-            white-space: nowrap;
-            overflow: hidden;
-            text-overflow: ellipsis;
-        }
+        white-space: nowrap;
+        overflow: hidden;
+        text-overflow: ellipsis;
+        // .title {
+        // }
     }
     .footer {
         grid-area: footer;
